Add new recipe page & extend index.html with a redirecting link
Recipes folder has been created. A new recipe webpage was created by
adding pesto-pasta-with-chicken.html inside that folder. Boilerplate
for that file was created & the file got populated with a head title
and a heading in body.

index.html was modified by adding an anchor, that is redirecting to
the new recipe pesto-pasta-with-chicken.html webpage.